added reservation model into livewire component
added reservation model into livewire component to start binding the data.

used mount method to mount it to the event.

added user as a public variable to take the authenticated user into that variable so that the user will have his own details.

defined public reservation variable to have a reservation object saved to start filling the object with details.